Get rid of IModuleDataHandler
Turns out the requirements are more or less the same as IPartModifier - different things need to respond to different events
