--- B9PartSwitch/PartSwitch/PartModifiers/IModuleDataHandler.cs
@@ -1,10 +1,0 @@
-﻿using System;
-
-namespace B9PartSwitch.PartSwitch.PartModifiers
-{
-    public interface IModuleDataHandler
-    {
-        void Activate();
-        void Deactivate();
-    }
-}
